Use valid thresholds config instead of invalid default_challenge key
The default_challenge key in botPolicies YAML was not a valid Anubis
config option and was being ignored. Changed template to generate
proper thresholds config that Anubis actually reads.

Also adds InSpec test infrastructure using osl-nginx for the reverse
proxy to verify the difficulty is correctly applied in the challenge
response.

Signed-off-by: Lance Albertson <lance@osuosl.org>